refactor: refactor ledger
