+++ packages/ledger/src/index.ts
@@ -1,3 +1,4 @@
 export * from './provider';
 export * from './types';
 export * from './ledger';
+export { type LedgerErrorCode, LedgerError } from './ledger/errors';
