+++ packages/ledger/src/ledger/errors.ts
@@ -1,0 +1,26 @@
+/**
+ * Ledger error codes
+ */
+export type LedgerErrorCode =
+  | 'NO_DEVICE'
+  | 'CONNECTION_FAILED'
+  | 'ETHEREUM_APP_NOT_OPEN'
+  | 'CANNOT_GET_ADDRESS'
+  | 'NO_SESSION'
+  | 'SIGN_MESSAGE_FAILED'
+  | 'SIGN_TYPED_DATA_FAILED';
+
+/**
+ * Extended Error class with code and details
+ */
+export class LedgerError extends Error {
+  code: LedgerErrorCode;
+  details?: string;
+
+  constructor(code: LedgerErrorCode, details: string) {
+    super(details);
+    this.code = code;
+    this.details = details;
+    this.name = 'LedgerError';
+  }
+}
